refactor : Plan, PlaceCard 경로 변경
