+++ src/routes/PlanningPage/Planning.css
@@ -17,6 +17,50 @@
   background-color: #f6f3df;
 }
 
+.div-left-title {
+  float: top;
+  width: 100%;
+  height: 8%;
+  background-color: #f5f3f4;
+  display: flex; /* Flexbox를 사용하여 수직 가운데 정렬 */
+  align-items: center; /* 수직 가운데 정렬 */
+  font-size: 24px; /* 글자 크기를 10px로 설정 */
+  padding: 20px;
+}
+
+.div-left-writer {
+  float: top;
+  width: 100%;
+  height: 3%;
+  background-color: #e6e8e6;
+}
+
+.div-plan-list {
+  display: flex;
+  flex-wrap: wrap;
+  align-content: flex-start;
+  width: 100%;
+  height: 82%;
+  background-color: #ffffff;
+  overflow: auto;
+  padding-left: 10px;
+}
+
+.div-plan-button {
+  width: 100%;
+  height: 7%;
+  background-color: #999999;
+}
+
+.titleText {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100px;
+  font-size: 10px;
+  margin: 0;
+}
+
 .toggle-right-up {
   position: absolute;
   right: 15px;
@@ -59,17 +103,21 @@
 .toggle-right-search-div3 {
   display: flex;
   flex-wrap: wrap;
+  align-content: flex-start;
   width: 100%;
   height: 75%;
+  padding-left: 13px;
   overflow: auto;
 }
 
 .toggle-right-basket-div1 {
   display: flex;
   flex-wrap: wrap;
+  align-content: flex-start;
   width: 100%;
   height: 90%;
   overflow: auto;
+  padding-left: 13px;
 }
 
 .toggle-right-basket-div2 {
@@ -110,33 +158,3 @@
   color: white;
   border-radius: 5px;
 }
-
-.div-left-title {
-  float: top;
-  width: 100%;
-  height: 8%;
-  background-color: #f5f3f4;
-}
-
-.div-left-writer {
-  float: top;
-  width: 100%;
-  height: 3%;
-  background-color: #e6e8e6;
-}
-
-.div-plan-list {
-  width: 100%;
-  height: 82%;
-  background-color: #ffffff;
-  overflow: auto;
-}
-
-.div-plan-button {
-  width: 100%;
-  height: 7%;
-  background-color: #999999;
-  overflow: auto;
-}
-
-/* absolute right-4 top-4 btn btn-sm btn-primary */
